feat: add SslResponse message
The backend should communicate with the frontend using backend messages.
But this simple concept is broken when you want to response to
SslRequest messsage. There is no such message as SslResponse in
postgres, so to answer to SslRequest they use a very special approach:
the backend sends a single byte containing 'S' or 'N',
indicating that it is willing or unwilling to perform SSL, respectively.

I don't like how this is done in postgres and, as for me, it is very
unnatural, so this patch introduces SslResponse message, that simply
wraps this single byte and provides the backend with convenient API
for responding to an SslRequest message.

Also it is better in terms of symmetry: there is SslRequest in frontend
messages, but there is no SslResponse in backend messages.