feat: respect `CopyInProgress` state
The recently introduced `CopyInProgress` state would be usually set by
the code that handles a `COPY <t> FROM STDIN` statement.

The `process_message` loop should respect this behavior and only act on
`CopyData` / `CopyDone` / `CopyFail` messages in this state.

In addition, the default `SimpleQueryHandler::on_query` implementation
should not transition back to a `ReadyForQuery` state or send a response
to the client if the `do_query` method ended with initiating a `COPY`
phase.

The code is now updated so the `⇒ CopyInProgress` transition happens in
the `on_query` default implementations, and the `⇒ ReadyForQuery`
transition happens in the `process_message` implementation.

The `CopyHandler` implementation in `examples/copy.rs` is also udpated
to reflect the asserted state.