feat: implement `ToSqlText` for `Decimal`
The `rust_decimal` crate provides a `Decimal` type that implements
the `ToSQL` trait required for integration with `rust-postgres`.

This is one of the recommended ways[^1] of exposing `NUMERIC` and
`DECIMAL` columns in `rust-postgres` / `tokio-postgres` results.

This commit introduces `rust_decimal` as an optional dependency and
adds a `ToSqlText` implementation (similar to how the `chrono` types
are pulled in and used to provide support for `DATE`/`TIME`/`DATETIME`
Postgres types).

[^1]: https://stackoverflow.com/a/72626605