fix: use u16 type for parameters len
According to the PostgreSQL documentation, the upper limit for query
parameters is 65,535, that cannot be represented as i16.
(https://www.postgresql.org/docs/16/limits.html)

The old implementation caused a panic for parse and bind messages with
more than 32,767 parameters. For instance, 65,535 turned into -1, which
when cast to usize resulted in usize::MAX. This led to a panic during
the allocation of a vector for usize::MAX parameters.